Add get id num from database when writiing to db, implemente add new menu item for dishes and bev, change all double val to string val for db compatability

+++ src/com/company/menu/SetMenu.java
@@ -4,9 +4,9 @@ import java.time.LocalDate;
 
 
 public class SetMenu extends Menu {
-    int set_menu_price = 0;
+    double set_menu_price = 0.0;
 
-    public SetMenu(int id, String name, String description, LocalDate date, int set_menu_price) {
+    public SetMenu(int id, String name, String description, LocalDate date, double set_menu_price) {
         super(id, name, description, date);
         this.set_menu_price = set_menu_price;
     }
